refactor: extract gateway pairing runtime helpers
